+++ src/navigation/AppNav.tsx
@@ -5,7 +5,6 @@ import { ROUTES } from './routes';
 import { AppNavbar } from '../components';
 import { About } from '../pages/about/About';
 import { MainLayout } from '../ui/MainLayout';
-import { QuadrantView } from '../pages/views';
 import { NotFound404, Radar, Search } from '../pages';
 
 export const NavApp = () => (
@@ -13,11 +12,14 @@ export const NavApp = () => (
     <AppNavbar />
     <MainLayout>
       <Routes>
-        <Route path={ROUTES.RADAR} element={<Radar />} />
+        <Route path={ROUTES.RADAR} element={<Radar />}>
+          <Route path={ROUTES.QUADRANT}>
+            <Route path={ROUTES.QUADRANT_PARAM} element={<></>} />
+          </Route>
+        </Route>
         <Route path={ROUTES.ABOUT} element={<About />} />
         <Route path={ROUTES.SEARCH} element={<Search />} />
 
-        <Route path={ROUTES.QUADRANT_PARAM} element={<QuadrantView />} />
         {/* https://gist.github.com/mjackson/b5748add2795ce7448a366ae8f8ae3bb#not-server-rendering -> should be server redirect */}
         <Route path='/' element={<Navigate replace to={ROUTES.RADAR} />} />
         <Route path='*' element={<NotFound404 />} />
