+++ src/pages/Radar.tsx
@@ -1,9 +1,6 @@
 import React, { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import {
-  // BlipType,
-  useRadarState
-} from '@undp_sdg_ai_lab/undp-radar';
+import { useNavigate, useParams } from 'react-router-dom';
+import { useRadarState } from '@undp_sdg_ai_lab/undp-radar';
 
 import { ROUTES } from '../navigation/routes';
 import { RadarView } from './views';
@@ -11,32 +8,33 @@ import { RadarView } from './views';
 export const Radar: React.FC = () => {
   const nav = useNavigate();
   const {
-    state: { selectedItem, selectedQuadrant },
-    actions: { setSelectedItem, setSelectedQuadrant }
+    state: {
+      selectedItem,
+      selectedQuadrant,
+      radarData: { quadrants }
+    },
+    actions: { setSelectedQuadrant }
   } = useRadarState();
 
-  useEffect(() => {
-    // return () => {
-    //   logic.setSelectedItem(null);
-    //   logic.setSelectedQuadrant(null);
-    // };
-  }, [setSelectedItem, setSelectedQuadrant]);
+  const goToQuadrant = (quadrant: string) =>
+    nav(`${ROUTES.QUADRANT}/${quadrant}`);
 
   useEffect(() => {
-    const goToQuadrant = (quadrant: string) =>
-      nav(`${ROUTES.QUADRANT}/${quadrant}`);
-    // const goToBlip = (blip: BlipType) => nav(`${ROUTES.BLIP}/${blip.id}`);
-    if (selectedItem) {
-      // go to Blip view
-      // goToBlip(selectedItem);
-    } else if (!selectedItem && selectedQuadrant) {
-      // go to quadrant view
-      goToQuadrant(selectedQuadrant);
-    } else {
-      // !selectedItem && !selectedQuadrant
-      // Pass through so we can see the radar
-    }
-  }, [selectedItem, selectedQuadrant, nav]);
+    if (!selectedItem && selectedQuadrant) goToQuadrant(selectedQuadrant);
+  }, [selectedItem, selectedQuadrant]);
+
+  const { quadrantId } = useParams();
+
+  useEffect(() => {
+    if (
+      !selectedItem &&
+      quadrantId &&
+      quadrants &&
+      quadrants.length > 0 &&
+      quadrants.includes(quadrantId)
+    )
+      setSelectedQuadrant(quadrantId);
+  }, [selectedItem, selectedQuadrant, quadrants, quadrantId]);
 
   return <RadarView />;
 };
