+++ src/pages/views/index.ts
@@ -1,4 +1,3 @@
-import { QuadrantView } from './QuadrantView';
 import { RadarView } from './RadarView';
 
-export { QuadrantView, RadarView };
+export { RadarView };
